fix: dt weight decay
