+++ offlinerllib/module/net/attention/dt.py
@@ -1,0 +1,71 @@
+from typing import Any, Dict, List, Optional, Sequence, Tuple, Type, Union
+
+import torch
+import torch.nn as nn
+
+from offlinerllib.module.net.attention.gpt2 import GPT2
+
+class DecisionTransformer(GPT2):
+    def __init__(
+        self, 
+        obs_dim: int, 
+        action_dim: int, 
+        embed_dim: int, 
+        num_layers: int, 
+        seq_len: int, 
+        episode_len: int=1000, 
+        num_heads: int=1, 
+        attention_dropout: Optional[float]=0.1, 
+        residual_dropout: Optional[float]=0.1, 
+        embed_dropout: Optional[float]=0.1, 
+    ) -> None:
+        super().__init__(
+            input_dim=embed_dim, # actually not used
+            embed_dim=embed_dim, 
+            num_layers=num_layers, 
+            seq_len=3*seq_len, # (return_to_go, state, action)
+            num_heads=num_heads, 
+            causal=True, 
+            attention_dropout=attention_dropout, 
+            residual_dropout=residual_dropout, 
+            embed_dropout=embed_dropout, 
+        )
+        # we manually do the embeddings here
+        self.pos_embed = nn.Embedding(episode_len + seq_len, embed_dim)
+        self.obs_embed = nn.Linear(obs_dim, embed_dim)
+        self.act_embed = nn.Linear(action_dim, embed_dim)
+        self.ret_embed = nn.Linear(1, embed_dim)
+        
+        self.action_head = nn.Sequential(nn.Linear(embed_dim, action_dim), nn.Tanh())
+
+    def forward(
+        self, 
+        states: torch.Tensor, 
+        actions: torch.Tensor, 
+        returns_to_go: torch.Tensor, 
+        timesteps: torch.Tensor, 
+        attention_mask: Optional[torch.Tensor]=None, 
+        key_padding_mask: Optional[torch.Tensor]=None, 
+    ):
+        B, L, *_ = states.shape
+        time_embedding = self.pos_embed(timesteps)
+        state_embedding = self.obs_embed(states) + time_embedding
+        action_embedding = self.act_embed(actions) + time_embedding
+        return_embedding = self.ret_embed(returns_to_go) + time_embedding
+        
+        if key_padding_mask is not None:
+            key_padding_mask = torch.stack([key_padding_mask, key_padding_mask, key_padding_mask], dim=2).reshape(B, 3*L)
+        
+        stacked_input = torch.stack([return_embedding, state_embedding, action_embedding], dim=2).reshape(B, 3*L, state_embedding.shape[-1])
+        out = super().forward(
+            inputs=stacked_input, 
+            timesteps=None, 
+            attention_mask=attention_mask, 
+            key_padding_mask=key_padding_mask, 
+            do_embedding=False
+        )
+
+        out = self.action_head(out[:, 1::3])
+        return out    # (batch size, length, action_shape)
+
+
