+++ offlinerllib/module/net/attention/gpt2.py
@@ -1,15 +1,10 @@
-from turtle import back
 from typing import Any, Dict, List, Optional, Sequence, Tuple, Type, Union
 
 import torch
 import torch.nn as nn
 
-from offlinerllib.module.net.basic import miniblock, EnsembleLinear
 
-Attention: nn.Module = nn.MultiheadAttention
-
-
-class TransformerBlock(nn.Module):
+class GPTBlock(nn.Module):
     def __init__(
         self, 
         embed_dim: int, 
@@ -23,12 +18,13 @@ class TransformerBlock(nn.Module):
         self.seq_len = seq_len
         if backbone_dim is None:
             backbone_dim = 4 * embed_dim
-        self.attention = Attention(
+        self.attention = nn.MultiheadAttention(
             embed_dim=embed_dim, 
             num_heads=num_heads, 
             dropout=attention_dropout, 
             batch_first=True
         )
+        self.drop = nn.Dropout(residual_dropout)
         self.ln1 = nn.LayerNorm(embed_dim)
         self.ln2 = nn.LayerNorm(embed_dim)
         self.ff = nn.Sequential(
@@ -60,12 +56,12 @@ class TransformerBlock(nn.Module):
             attn_mask=attention_mask, 
             key_padding_mask=key_padding_mask
         )[0]
-        residual = residual + attn_output  # check: we escape the dropout, original no dropout, corl does dropout although i don't think that is right
+        residual = residual + self.drop(attn_output) # this is because pytorch MHV don't do dropout after final projection
         residual = residual + self.ff(self.ln2(residual))
         return residual
         
         
-class Transformer(nn.Module):
+class GPT2(nn.Module):
     def __init__(
         self, 
         input_dim: int, 
@@ -74,9 +70,9 @@ class Transformer(nn.Module):
         seq_len: int, 
         num_heads: int, 
         causal: bool=False, 
-        attention_dropout: Optional[float]=None, 
-        residual_dropout: Optional[float]=None, 
-        embed_dropout: Optional[float]=None, 
+        attention_dropout: Optional[float]=0.1, 
+        residual_dropout: Optional[float]=0.1, 
+        embed_dropout: Optional[float]=0.1, 
         position_dim: Optional[int]=None
     ) -> None:
         super().__init__()
@@ -85,9 +81,8 @@ class Transformer(nn.Module):
         self.pos_embed = nn.Embedding(position_dim, embed_dim)
         self.embed_dropout = nn.Dropout(embed_dropout) if embed_dropout else nn.Identity()
         self.out_ln = nn.LayerNorm(embed_dim)
-        # self.embed_ln = nn.LayerNorm(embed_dim)   # check: whether or not add layer norm before embed dropout
         self.blocks = nn.ModuleList([
-            TransformerBlock(
+            GPTBlock(
                 embed_dim=embed_dim, 
                 seq_len=seq_len, 
                 num_heads=num_heads, 
@@ -127,68 +122,37 @@ class Transformer(nn.Module):
         inputs = self.out_ln(inputs)
         return inputs
     
-    
-class DecisionTransformer(Transformer):
-    def __init__(
-        self, 
-        obs_dim: int, 
-        action_dim: int, 
-        embed_dim: int, 
-        num_layers: int, 
-        seq_len: int, 
-        episode_len: int=1000, 
-        num_heads: int=4, 
-        attention_dropout: Optional[float]=None, 
-        residual_dropout: Optional[float]=None, 
-        embed_dropout: Optional[float]=None, 
-    ) -> None:
-        super().__init__(
-            input_dim=embed_dim, # actually not used
-            embed_dim=embed_dim, 
-            num_layers=num_layers, 
-            seq_len=3*seq_len, # (return_to_go, state, action)
-            num_heads=num_heads, 
-            causal=True, 
-            attention_dropout=attention_dropout, 
-            residual_dropout=residual_dropout, 
-            embed_dropout=embed_dropout, 
-        )
-        # we manually do the embeddings here
-        self.pos_embed = nn.Embedding(episode_len + seq_len, embed_dim)
-        self.obs_embed = nn.Linear(obs_dim, embed_dim)
-        self.act_embed = nn.Linear(action_dim, embed_dim)
-        self.ret_embed = nn.Linear(1, embed_dim)
-        
-        self.action_head = nn.Sequential(nn.Linear(embed_dim, action_dim), nn.Tanh())
+    def configure_params(self):
+        # https://github.com/karpathy/minGPT/blob/37baab71b9abea1b76ab957409a1cc2fbfba8a26/mingpt/model.py#L215
+        # However in this method we are not returning an Optimizer, but the parameter groups which
+        # need / needn't weight decay, to support for more flexible downstream application
+        decay = set()
+        no_decay = set()
+        whitelist_weight_modules = (torch.nn.Linear, torch.nn.MultiheadAttention)
+        blacklist_weight_modules = (torch.nn.LayerNorm, torch.nn.Embedding)
+        for mn, m in self.named_modules():
+            for pn, p in m.named_parameters():
+                fpn = '%s.%s' % (mn, pn) if mn else pn # full param name
+                # random note: because named_modules and named_parameters are recursive
+                # we will see the same tensors p many many times. but doing it this way
+                # allows us to know which parent module any tensor p belongs to...
+                if pn.endswith('bias'):
+                    # all biases will not be decayed
+                    no_decay.add(fpn)
+                elif pn.endswith('weight') and isinstance(m, whitelist_weight_modules):
+                    # weights of whitelist modules will be weight decayed
+                    decay.add(fpn)
+                elif pn.endswith('weight') and isinstance(m, blacklist_weight_modules):
+                    # weights of blacklist modules will NOT be weight decayed
+                    no_decay.add(fpn)
+            
+        # validate that we considered every parameter
+        param_dict = {pn: p for pn, p in self.named_parameters()}
+        inter_params = decay & no_decay
+        union_params = decay | no_decay
+        assert len(inter_params) == 0, "parameters %s made it into both decay/no_decay sets!" % (str(inter_params), )
+        assert len(param_dict.keys() - union_params) == 0, "parameters %s were not separated into either decay/no_decay set!" \
+                                                    % (str(param_dict.keys() - union_params), )
 
-    def forward(
-        self, 
-        states: torch.Tensor, 
-        actions: torch.Tensor, 
-        returns_to_go: torch.Tensor, 
-        timesteps: torch.Tensor, 
-        attention_mask: Optional[torch.Tensor]=None, 
-        key_padding_mask: Optional[torch.Tensor]=None, 
-    ):
-        B, L, *_ = states.shape
-        time_embedding = self.pos_embed(timesteps)
-        state_embedding = self.obs_embed(states) + time_embedding
-        action_embedding = self.act_embed(actions) + time_embedding
-        return_embedding = self.ret_embed(returns_to_go) + time_embedding
-        
-        if key_padding_mask is not None:
-            key_padding_mask = torch.stack([key_padding_mask, key_padding_mask, key_padding_mask], dim=2).reshape(B, 3*L)
-        
-        stacked_input = torch.stack([return_embedding, state_embedding, action_embedding], dim=2).reshape(B, 3*L, state_embedding.shape[-1])
-        out = super().forward(
-            inputs=stacked_input, 
-            timesteps=None, 
-            attention_mask=attention_mask, 
-            key_padding_mask=key_padding_mask, 
-            do_embedding=False
-        )
+        return [param_dict[pn] for pn in sorted(list(decay))], [param_dict[pn] for pn in sorted(list(no_decay))]
 
-        out = self.action_head(out[:, 1::3])
-        return out    # (batch size, length, action_shape)
-
-
